# tracing: simplify conditional redaction for tenant traces