--- a/pkg/server/node.go
+++ b/pkg/server/node.go
@@ -1057,10 +1057,17 @@ func (n *Node) setupSpanForIncomingRPC(
 	// remoteTrace case below.
 	const opName = "/cockroach.roachpb.Internal/Batch"
 	tr := n.storeCfg.AmbientCtx.Tracer
+
+	// We require redactability enabled on tenant requests.
+	// TODO(obs-inf): Once performance issues around redaction are
+	// resolved via #58610, this code can be removed so that all traces
+	// have redactability enabled.
+	forceRedactable := tenID != roachpb.SystemTenantID
+
 	var newSpan, grpcSpan *tracing.Span
 	if isLocalRequest := grpcutil.IsLocalRequestContext(ctx) && tenID == roachpb.SystemTenantID; isLocalRequest {
 		// This is a local request which circumvented gRPC. Start a span now.
-		ctx, newSpan = tracing.EnsureChildSpan(ctx, tr, opName)
+		ctx, newSpan = tracing.EnsureChildSpan(ctx, tr, opName, tracing.WithForceRedactable(forceRedactable))
 		// Set the same span.kind tag as the gRPC interceptor.
 		newSpan.SetTag(ext.SpanKindRPCServer.Key, ext.SpanKindRPCServer.Value)
 	} else {
@@ -1069,8 +1076,9 @@ func (n *Node) setupSpanForIncomingRPC(
 			// If tracing information was passed via gRPC metadata, the gRPC interceptor
 			// should have opened a span for us. If not, open a span now (if tracing is
 			// disabled, this will be a noop span).
-			ctx, newSpan = tr.StartSpanCtx(ctx, opName)
+			ctx, newSpan = tr.StartSpanCtx(ctx, opName, tracing.WithForceRedactable(forceRedactable))
 		} else {
+			grpcSpan.SetRedactable(forceRedactable)
 			grpcSpan.SetTag("node", n.Descriptor.NodeID)
 		}
 	}