--- a/pkg/server/status.go
+++ b/pkg/server/status.go
@@ -670,7 +670,7 @@ func recordedSpansToTraceEvents(spans []tracingpb.RecordedSpan) []*serverpb.Trac
 				Time: entry.Time,
 			}
 			if len(entry.Fields) == 1 {
-				event.Message = entry.Fields[0].Value.StripMarkers()
+				event.Message = entry.Fields[0].Value.StripMarkers(sp.RedactableLogs)
 			} else {
 				buf.Reset()
 				for i, f := range entry.Fields {