--- a/pkg/sql/exec_util.go
+++ b/pkg/sql/exec_util.go
@@ -2467,7 +2467,7 @@ func getMessagesForSubtrace(
 			allLogs = append(allLogs,
 				logRecordRow{
 					timestamp: logTime,
-					msg:       span.Logs[i].Msg(),
+					msg:       span.Logs[i].Msg(span.RedactableLogs),
 					span:      span,
 					// Add 1 to the index to account for the first dummy message in a
 					// span.