--- a/pkg/util/log/channels.go
+++ b/pkg/util/log/channels.go
@@ -53,7 +53,9 @@ func logfDepth(
 		ctx, sev, ch,
 		depth+1, true /* redactable */, format, args...)
 	if sp, el, ok := getSpanOrEventLog(ctx); ok {
-		eventInternal(sp, el, (sev >= severity.ERROR), entry.convertToLegacy())
+		// Prevent `entry` from moving to the heap if this branch isn't taken.
+		heapEntry := entry
+		eventInternal(sp, el, sev >= severity.ERROR, &heapEntry)
 	}
 	logger.outputLogEntry(entry)
 }