… pkg/ccl/kvccl/kvfollowerreadsccl/followerreads_test.go
… pkg/ccl/kvccl/kvtenantccl/tenant_trace_test.go
… pkg/kv/bulk/sst_batcher_test.go
… pkg/kv/kvserver/concurrency/concurrency_manager_test.go
… pkg/server/BUILD.bazel
… pkg/server/node_tenant_test.go
… pkg/server/node_test.go
… pkg/sql/opt/exec/explain/plan_gist_factory.og.go
… pkg/sql/rowexec/tablereader_test.go
… pkg/util/log/clog_test.go
… pkg/util/log/trace_test.go
… pkg/util/tracing/recording.go
… pkg/util/tracing/span.go
… pkg/util/tracing/span_inner.go
… pkg/util/tracing/span_options.go
… pkg/util/tracing/span_test.go
… pkg/util/tracing/test_utils.go
… pkg/util/tracing/tracer.go
… pkg/util/tracing/tracingpb/BUILD.bazel
… pkg/util/tracing/tracingpb/recorded_span.pb.go
… pkg/util/tracing/tracingpb/recorded_span.proto