This PR simplifies the proposed redaction model for tenant-bound traces.
It removes the need for a flag on recorded spans by simply retaining the
`RedactableString` type on the log field values and coarsely redacting
them in the common case manually by surrounding the message with
redaction tags, and only truly redacting it field-by-field if the span
has been marked redactable.

This approach removes a key complication of prior work in conditional
redaction by eliminating any changes to protos and eliminating the need
for a custom `MaybeRedactableString` type. Benchmarks support this
approach as not significantly worse in performance from the prior
commit.

----
Benchmarks

The base for all 3 below benchmarks that's used is the code on
`release-21.2` with the "drop traces for tenants" commit reverted. This
reproduces a reasonable "base" state where tracing redaction isn't
really introduced at all.

Additionally, all benchmarks are run with modifications introduced in
the first two commits in #71610 which force tracing on in benchmarks by
running `SET CLUSTER SETTING trace.debug.enable = true` in the regular
and multinode scenarios shown below.

Below is the benchmark for the original conditional redaction PR prior
to changes in this commit.
```
19:36:36 ❯ benchstat 83fc45287f-no-tracing-bench.txt cae9e1c773-conditional-redaction2.txt
name                                           old time/op    new time/op    delta
Tracing/Cockroach/tracing=f/Scan1-8               612µs ± 9%     598µs ± 2%     ~     (p=0.436 n=10+10)
Tracing/Cockroach/tracing=f/Insert-8              785µs ± 4%     761µs ± 3%   -3.07%  (p=0.002 n=8+9)
Tracing/Cockroach/tracing=t/Scan1-8               597µs ± 1%     582µs ± 1%   -2.59%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              762µs ± 3%     767µs ± 4%     ~     (p=0.315 n=10+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8     1.05ms ±30%    0.95ms ± 2%   -9.19%  (p=0.000 n=9+10)
Tracing/MultinodeCockroach/tracing=f/Insert-8    1.19ms ±15%    1.18ms ± 5%     ~     (p=0.243 n=10+9)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      979µs ± 3%     963µs ± 5%     ~     (p=0.297 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8    1.54ms ±48%    1.22ms ±11%  -21.19%  (p=0.023 n=10+10)

name                                           old alloc/op   new alloc/op   delta
Tracing/Cockroach/tracing=f/Scan1-8              86.5kB ± 1%    94.0kB ± 1%   +8.71%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=f/Insert-8              109kB ± 3%     117kB ± 3%   +7.86%  (p=0.000 n=9+8)
Tracing/Cockroach/tracing=t/Scan1-8              86.0kB ± 1%    94.3kB ± 2%   +9.66%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              108kB ± 3%     118kB ± 1%   +8.46%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      156kB ±37%     155kB ± 1%     ~     (p=0.059 n=9+8)
Tracing/MultinodeCockroach/tracing=f/Insert-8     154kB ±10%     165kB ± 7%   +7.31%  (p=0.005 n=8+8)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      145kB ± 1%     155kB ± 1%   +6.67%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     157kB ±13%     189kB ±34%  +20.64%  (p=0.006 n=8+10)

name                                           old allocs/op  new allocs/op  delta
Tracing/Cockroach/tracing=f/Scan1-8               1.44k ± 0%     1.48k ± 0%   +3.18%  (p=0.000 n=9+7)
Tracing/Cockroach/tracing=f/Insert-8              1.64k ± 0%     1.70k ± 0%   +3.35%  (p=0.000 n=9+8)
Tracing/Cockroach/tracing=t/Scan1-8               1.44k ± 0%     1.48k ± 0%   +2.94%  (p=0.000 n=8+8)
Tracing/Cockroach/tracing=t/Insert-8              1.64k ± 0%     1.70k ± 0%   +3.41%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      2.41k ±11%     2.37k ± 1%     ~     (p=0.156 n=10+9)
Tracing/MultinodeCockroach/tracing=f/Insert-8     2.15k ± 1%     2.21k ± 1%   +2.81%  (p=0.000 n=8+8)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      2.31k ± 1%     2.36k ± 0%   +2.33%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     2.17k ± 6%     2.33k ±10%   +7.30%  (p=0.006 n=8+10)
```

Below are tracing benchmarks for the simpler conditional redaction that
is in **this** commit. It appears to have a similar performance hit to
the one above with a simpler diff and no protobuf changes.
```
~/go/src/github.com/cockroachdb/cockroach make-traces-redactable-and-default-to-off--v2* ≡
19:36:43 ❯ benchstat 83fc45287f-no-tracing-bench.txt 2bea5a2ac5-conditional-redaction-simple.txt
name                                           old time/op    new time/op    delta
Tracing/Cockroach/tracing=f/Scan1-8               612µs ± 9%     591µs ±16%     ~     (p=0.393 n=10+10)
Tracing/Cockroach/tracing=f/Insert-8              785µs ± 4%     767µs ± 6%     ~     (p=0.360 n=8+10)
Tracing/Cockroach/tracing=t/Scan1-8               597µs ± 1%     542µs ± 1%   -9.23%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              762µs ± 3%     739µs ± 8%     ~     (p=0.315 n=10+10)
Tracing/MultinodeCockroach/tracing=f/Scan1-8     1.05ms ±30%    0.92ms ± 5%  -12.57%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Insert-8    1.19ms ±15%    1.13ms ± 9%   -5.39%  (p=0.035 n=10+10)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      979µs ± 3%     945µs ± 6%   -3.42%  (p=0.024 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8    1.54ms ±48%    1.19ms ±16%  -22.76%  (p=0.011 n=10+10)

name                                           old alloc/op   new alloc/op   delta
Tracing/Cockroach/tracing=f/Scan1-8              86.5kB ± 1%    90.2kB ± 3%   +4.29%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=f/Insert-8              109kB ± 3%     111kB ± 0%   +2.68%  (p=0.005 n=9+7)
Tracing/Cockroach/tracing=t/Scan1-8              86.0kB ± 1%    89.8kB ± 0%   +4.42%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              108kB ± 3%     112kB ± 3%   +3.05%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      156kB ±37%     149kB ± 1%     ~     (p=0.050 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Insert-8     154kB ±10%     154kB ± 3%     ~     (p=0.382 n=8+8)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      145kB ± 1%     150kB ± 5%   +3.61%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     157kB ±13%     183kB ±40%     ~     (p=0.237 n=8+10)

name                                           old allocs/op  new allocs/op  delta
Tracing/Cockroach/tracing=f/Scan1-8               1.44k ± 0%     1.50k ± 0%   +4.22%  (p=0.000 n=9+8)
Tracing/Cockroach/tracing=f/Insert-8              1.64k ± 0%     1.72k ± 0%   +4.39%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=t/Scan1-8               1.44k ± 0%     1.50k ± 0%   +4.33%  (p=0.000 n=8+7)
Tracing/Cockroach/tracing=t/Insert-8              1.64k ± 0%     1.72k ± 0%   +4.34%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      2.41k ±11%     2.38k ± 1%     ~     (p=0.150 n=10+9)
Tracing/MultinodeCockroach/tracing=f/Insert-8     2.15k ± 1%     2.22k ± 1%   +3.35%  (p=0.000 n=8+8)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      2.31k ± 1%     2.40k ± 4%   +4.19%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     2.17k ± 6%     2.33k ±12%   +7.43%  (p=0.012 n=8+10)
```

These statistics are the *base* hit we take on the redaction work
**prior** to any conditional redaction being added. This is what we're
trying to improve on.
```
~/go/src/github.com/cockroachdb/cockroach make-traces-redactable-and-default-to-off--v2* ≡
19:36:49 ❯ benchstat 83fc45287f-no-tracing-bench.txt dc42e3762d-after-redaction.txt
name                                           old time/op    new time/op    delta
Tracing/Cockroach/tracing=f/Scan1-8               612µs ± 9%     693µs ± 2%  +13.22%  (p=0.000 n=10+10)
Tracing/Cockroach/tracing=f/Insert-8              785µs ± 4%     878µs ± 1%  +11.86%  (p=0.000 n=8+9)
Tracing/Cockroach/tracing=t/Scan1-8               597µs ± 1%     687µs ± 3%  +15.05%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              762µs ± 3%     878µs ± 1%  +15.35%  (p=0.000 n=10+8)
Tracing/MultinodeCockroach/tracing=f/Scan1-8     1.05ms ±30%    1.06ms ± 2%     ~     (p=0.050 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Insert-8    1.19ms ±15%    1.28ms ±11%   +7.85%  (p=0.035 n=10+10)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      979µs ± 3%    1066µs ± 1%   +8.87%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8    1.54ms ±48%    1.26ms ± 2%     ~     (p=0.515 n=10+8)

name                                           old alloc/op   new alloc/op   delta
Tracing/Cockroach/tracing=f/Scan1-8              86.5kB ± 1%    98.0kB ± 1%  +13.33%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=f/Insert-8              109kB ± 3%     125kB ± 1%  +15.09%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=t/Scan1-8              86.0kB ± 1%    97.9kB ± 1%  +13.90%  (p=0.000 n=10+9)
Tracing/Cockroach/tracing=t/Insert-8              108kB ± 3%     125kB ± 1%  +15.53%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      156kB ±37%     162kB ± 1%     ~     (p=0.059 n=9+8)
Tracing/MultinodeCockroach/tracing=f/Insert-8     154kB ±10%     197kB ±43%  +28.54%  (p=0.000 n=8+10)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      145kB ± 1%     162kB ± 1%  +11.81%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     157kB ±13%     172kB ± 5%   +9.84%  (p=0.007 n=8+8)

name                                           old allocs/op  new allocs/op  delta
Tracing/Cockroach/tracing=f/Scan1-8               1.44k ± 0%     1.66k ± 0%  +15.59%  (p=0.000 n=9+9)
Tracing/Cockroach/tracing=f/Insert-8              1.64k ± 0%     1.95k ± 0%  +18.48%  (p=0.000 n=9+8)
Tracing/Cockroach/tracing=t/Scan1-8               1.44k ± 0%     1.66k ± 0%  +15.46%  (p=0.000 n=8+9)
Tracing/Cockroach/tracing=t/Insert-8              1.64k ± 0%     1.95k ± 1%  +18.52%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=f/Scan1-8      2.41k ±11%     2.58k ± 1%     ~     (p=0.105 n=10+8)
Tracing/MultinodeCockroach/tracing=f/Insert-8     2.15k ± 1%     2.58k ±12%  +20.04%  (p=0.000 n=8+10)
Tracing/MultinodeCockroach/tracing=t/Scan1-8      2.31k ± 1%     2.59k ± 1%  +12.09%  (p=0.000 n=9+9)
Tracing/MultinodeCockroach/tracing=t/Insert-8     2.17k ± 6%     2.47k ± 1%  +13.76%  (p=0.000 n=8+8)
```

Release note: None